requirements.txt
# Core ML/RL frameworks - GPU/CPU optimized
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0
torch>=1.13.0  # Auto-detects CUDA
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.8.0
scikit-learn>=1.1.0
gymnasium>=0.26.0

# Trading & Finance
kiteconnect>=4.0.0
pyotp>=2.6.0

# Technical Analysis  
TA-Lib>=0.4.25

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web automation
selenium>=4.0.0

# Data & HTTP
requests>=2.28.0
pytz>=2022.1
joblib>=1.2.0
multiprocessing-logging>=0.3.4

# System optimization packages
psutil>=5.8.0
tqdm>=4.64.0

# GPU acceleration with JAX/SBX
jax[cuda12_pip]>=0.4.0  # JAX with CUDA 12 support
sbx-rl>=0.12.0  # GPU-accelerated RL
optax>=0.1.7
flax>=0.7.0

# Optional ONNX support
onnx>=1.12.0
onnxruntime-gpu>=1.12.0  # GPU support for ONNX